Refactor dist tests
